requirements.txt
PyMuPDF==1.25.5
python-dotenv==1.1.0
